Aggain, lots of stuff changed. I have to get used to commit more frequently.
the interpolation function has been cleaned and debugged. Now it also corrects the start time of the trace, but is still not perfect.
Exploiting the antenna positions and the incoming direction for the intermediate interpolation steps should improve things, but since the intermediate positions are computed in the
antenna frame, is a little bit cumbersome to pass again to the ground plane to get the actual positions. ToDO in the future, if needed, i guess.

The interpolator now saves the P2P amplitudes and Hilbert peak value and time in the hdf5 file (in DEVELOPEMENT mode) of the last 16 antennas, for comparison.